Laurens Goedel Contribution
A little while ago we suggested to add a exit() around line 1183 on the
file /plugins/system/fieldsattachment/fieldsattachment.php.
For the frontend editing it’s also needed to add an exit() on line 1208
after the script document redirect.
This prevents in creating empty articles when the user cancel to create
a new article.