Fix for front end, list view, add button
This is a fix for an issue with the front end add button.  The issue is that when an article is created in the front end, from the add button in the list view, the category is lost on article creation.

To recreate this issue...
1) Take the standard Joomla install with sample data
2) login to the front-end as an administrator
3) navigate to a list view in then front end
4) click on the "add item" button
5) enter content in the new article
6) notice the category field has not been set, in the form, to the same category as the list view

This is quite a subtle issue really, and will only effect users who create articles from the front end.