Minor change to message wording, improved english
Hi Cristian

This is a small pull request, I hope this is useful.  It's just really to improve the working of save messages.

This is not a criticism of the wording, as it is understandable, but just to suggest perhaps better English wording.
"Uploda image Error" -> "Image upload error"
"Uploda image OK" -> "Image uploaded OK"

I hope that's ok to suggest.  fieldsattach is a fantastic system, so I feel bad for highlighting very minor things like this, but I only do so to try and help.

Thanks

Andy